Rebase over main and split large WIP puli 0.2.0 commit

+++ src/test/java/org/liveontologies/puli/RandomProofNode.java
@@ -36,8 +36,10 @@ public class RandomProofNode {
 		BaseProofBuilder<Integer, ?> b = new BaseProofBuilder<>();
 		for (int i = 0; i < maxInferences; i++) {
 			int conclusion = random.nextInt(maxConclusions);
-			  ProofBuilder<Integer, ?, ?>.ProofInferenceBuilder ib = b
-					.conclusion(conclusion);
+
+			 ProofBuilder<Integer, ?, ?>.ProofInferenceBuilder ib = b
+					 .conclusion((Integer) conclusion);
+
 			int noPremises = random.nextInt(maxPremises);
 			if (derived.size() < noPremises) {
 				noPremises = derived.size();
